Expose timeout configuration through Client
As a last step, we add the `set_timeouts` class method that returns
options for configuring a client with custom timeouts.

Both timeouts can be set at once, but it's also possible to configure
just one of them if necessary.

The `set_timeouts` method uses kwargs to explicitly define its
interface, whereas `timeout=` is a private method so it works with
hash arguments for convenience.