Set Net::HTTP timeouts in Request#execute
Timeout configuration is a client concern, but currently the
configuration is buried deep in internals of the library. So, as a first
step, we can move the configuration of the Request to the `#execute`
method.  We can leverage kwargs to provide default values for timeouts
and make them configurable from outside at the same time.

This change doesn't have tests because it would require dangereous
mocking and stubbing of objects we don't control.